renames plus three-node-cassandra test suite
