+++ atlasdb-ete-tests/docker-compose.single-client-with-embedded-and-three-node-cassandra.yml
@@ -30,17 +30,8 @@ services:
 
   ete1:
     build: .
-    command: bash -c 'cp var/conf/atlasdb-ete.multiple-cassandra.yml var/conf/atlasdb-ete.yml && dockerize -timeout 120s -wait tcp://cassandra1:9160 -wait tcp://cassandra2:9160 -wait tcp://cassandra3:9160 && service/bin/init.sh console'
+    command: bash -c 'cp var/conf/atlasdb-ete.embedded-and-three-node-cassandra.yml var/conf/atlasdb-ete.yml && dockerize -timeout 120s -wait tcp://cassandra1:9160 -wait tcp://cassandra2:9160 -wait tcp://cassandra3:9160 && service/bin/init.sh console'
     environment:
       - ME=ete1
     ports:
       - "3828"
-
-  ete-cli:
-    build: .
-    entrypoint: [bash, -c, 'cp var/conf/atlasdb-ete.multiple-cassandra.yml var/conf/atlasdb-ete.yml && "$$@"', --]
-    command: exit
-    environment:
-      - ME=ete1
-    depends_on:
-      - ete1
