+++ atlasdb-ete-tests/docker/conf/atlasdb-ete.embedded-and-three-node-cassandra.yml
@@ -29,15 +29,3 @@ atlasdb:
     replicationFactor: 3
 
   namespace: atlasete
-
-  leader:
-    quorumSize: 1
-    learnerLogDir: var/data/paxosLog/learner1
-    acceptorLogDir: var/data/paxosLog/acceptor1
-    localServer: http://${ME}:3828
-    leaders:
-      - http://ete1:3828
-    sslConfiguration:
-      keyStorePath: "var/security/keyStore.jks"
-      keyStorePassword: "keystore"
-      trustStorePath: "var/security/trustStore.jks"
