+++ atlasdb-ete-tests/src/test/java/com/palantir/atlasdb/ete/suites/SingleClientWithEmbeddedAndCassandraTestSuite.java
