+++ atlasdb-ete-tests/src/test/java/com/palantir/atlasdb/ete/suites/SingleClientWithEmbeddedAndPostgresTestSuite.java
@@ -16,9 +16,7 @@
 package com.palantir.atlasdb.ete;
 
 import com.google.common.collect.ImmutableList;
-import com.palantir.atlasdb.containers.CassandraEnvironment;
 import com.palantir.atlasdb.ete.coordination.CoordinationEteTest;
-import com.palantir.atlasdb.ete.coordination.MultipleSchemaVersionsCoordinationEteTest;
 import org.junit.ClassRule;
 import org.junit.rules.RuleChain;
 import org.junit.runner.RunWith;
@@ -27,15 +25,14 @@ import org.junit.runners.Suite;
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
     TodoEteTest.class,
-    ServiceExposureEteTest.class,
     TimestampManagementEteTest.class,
     CoordinationEteTest.class,
-    MultipleSchemaVersionsCoordinationEteTest.class
+    LockWithoutTimelockEteTest.class
 })
-public class CassandraMultinodeTestSuite extends EteSetup {
+public class SingleClientWithEmbeddedAndPostgres extends EteSetup {
     private static final ImmutableList<String> CLIENTS = ImmutableList.of("ete1", "ete2", "ete3");
 
     @ClassRule
-    public static final RuleChain COMPOSITION_SETUP = EteSetup.setupComposition(
-            CassandraMultinodeTestSuite.class, "docker-compose.cassandra.yml", CLIENTS, CassandraEnvironment.get());
+    public static final RuleChain COMPOSITION_SETUP =
+            EteSetup.setupComposition(SingleClientWithEmbeddedAndPostgres.class, "docker-compose.dbkvs.yml", CLIENTS);
 }
