+++ atlasdb-ete-tests/src/test/java/com/palantir/atlasdb/ete/suites/TestSuites.java
@@ -1,0 +1,2 @@
+package com.palantir.atlasdb.ete.suites;public class TestSuites {
+}
